api/requirements.txt
# API Dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
python-dotenv==1.0.1

# ML & Data
mlflow==2.18.0
xgboost==2.1.3
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.0
joblib==1.4.2

# Utilities
pyyaml==6.0.2
